add Exception for user

+++ src/main/java/com/devsync/service/UserService.java
@@ -2,6 +2,9 @@ package com.devsync.service;
 
 import com.devsync.dao.UserDao;
 import com.devsync.domain.entities.User;
+import com.devsync.exeptions.EmailExistException;
+import com.devsync.exeptions.InvalidArgumentException;
+import com.devsync.exeptions.ResourceNotFoundException;
 import com.devsync.scheduled.TaskRequestUpdater;
 import com.devsync.service.interfaces.UserServiceInterface;
 
@@ -13,8 +16,6 @@ public class UserService implements UserServiceInterface {
     private UserDao userDao;
     public UserService() {
         userDao = new UserDao();
-
-
     }
 
     @Override
@@ -24,21 +25,38 @@ public class UserService implements UserServiceInterface {
 
     @Override
     public User findById(Long userId) {
-        return userDao.findById(userId);
+        if (userId == null) {
+            throw new InvalidArgumentException("User ID cannot be null");
+        }
+        return userDao.findById(userId).orElseThrow(() ->
+                new ResourceNotFoundException("User with ID " + userId + " not found"));
     }
 
     @Override
     public User save(User user){
-        return  userDao.save(user);
+        if (user == null) {
+            throw new InvalidArgumentException("User cannot be null");
+        }
+            User existingUser = findByEmail(user.getEmail());
+            if (existingUser != null) {
+                throw new EmailExistException("This email is already used");
+            }
+            return userDao.save(user);
     }
 
     @Override
     public User update(User user){
-        return userDao.update(user);
+        if (user == null) {
+            throw new InvalidArgumentException("User cannot be null");
+        }
+            return userDao.update(user);
     }
 
     @Override
     public void delete(Long userId){
+        if (userId == null) {
+            throw new InvalidArgumentException("User id  cannot be null");
+        }
        userDao.delete(userId);
     }
 
@@ -49,6 +67,9 @@ public class UserService implements UserServiceInterface {
 
     @Override
     public User findByEmail(String email) {
+        if (email == null) {
+            throw new InvalidArgumentException("User Email cannot be null");
+        }
             return  userDao.findByEmail(email);
     }
 }
